Moved paddingIfNotNull to Modifiers extension file
