+++ liveview-android/src/main/java/org/phoenixframework/liveview/ui/phx_components/PhxLiveView.kt
@@ -1,10 +1,8 @@
 package org.phoenixframework.liveview.ui.phx_components
 
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
 import org.phoenixframework.liveview.domain.base.PushEvent
 import org.phoenixframework.liveview.domain.factory.ComposableNodeFactory
 import org.phoenixframework.liveview.domain.factory.ComposableTreeNode
@@ -28,10 +26,3 @@ fun PhxLiveView(
         paddingValues = paddingValues,
     )
 }
-
-fun Modifier.paddingIfNotNull(paddingValues: PaddingValues?): Modifier =
-    if (paddingValues != null) {
-        this.padding(paddingValues)
-    } else {
-        this
-    }
